Split metrics implementation into modules
